+++ docs/changelog.rst
@@ -14,6 +14,9 @@ Changelog
   using standard keyboards) via `IBus
   <https://github.com/ibus/ibus/wiki/ReadMe>`_ (:iss:`469`)
 
+- Implement completion for the kitty command in bash and zsh. See
+  :ref:`completion`.
+
 - Render the text under the cursor in a fixed color, configurable via
   the option :opt:`cursor_text_color` (:iss:`126`)